image_pipeline/requirements.txt
neo4j>=5.0,<6.0
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
numpy==2.1.3
opencv-python==4.10.0.84
python-multipart==0.0.19
svgwrite==1.4.3
ultralytics==8.3.58
easyocr>=1.7,<2.0
open_clip_torch>=2.24,<3.0
Pillow>=10,<12
